Update NUMBER.MaxRateValue.md
The decimal separator used is actually a comma not a point. The code handles an input using a point as the separator and changes e.g. 0.05 to 0,05. I suggest adding an example for absolute and relative changes, probably 0,05 would mean a 5% change if set to relative? What would be the use case for setting the MaxRateType to relative?